utils: Add web specific GstTask and GstTaskPool
This is required to support transferring objects to/from streaming
threads. We can't manipulate the life of a streaming thread but
we can override its task. This task/taskpool is special as it allows
Emscripten to process the WebWorker message handling while paused
and resumed properly afterwards.

Originally-from: gst.wasm