gobject-introspection.wrap: Backport patch to fix build with 'new' setuptools
This allows it to build with the newer glib subproject.
We can't update to the latest gobject-introspection yet because pygobject,
pycairo, gst-python all fail to build with it, and will also need updates.
There's a lot of breaking changes in the latest gobject-introspection.

Same logic as https://gitlab.freedesktop.org/gstreamer/gstreamer/-/merge_requests/8363