audioaggregator: support seeks and silenced reverse playback
In audio+video pipelines, the seek might travel upstream only in the video branch,
and we later get the updated segment events on the sink pads. We should
update the srcpad segment in that case.

Moreover, instead of outright rejecting seeks with negative rates, let's
just skip the aggregating and output silence.