gstinfo: Add convenience macros around GstLogContext for logging once
The purpose is to avoid having to add boilerplate to every file/plugin
where you want to add a log message which will be printed exactly
once, with the default settings for GstLogContext.

Somehow, this is also faster than using the explicit API:
GST_WARNING_ONCE is 15% faster than GST_CTX_WARNING

The macros are the only publicly-documented API, the functions that
are wrapped by the macros are not meant to be used directly, except by
bindings.

Part-of: <https://gitlab.freedesktop.org/gstreamer/gstreamer/-/merge_requests/9939>